Fix Ethos-U driver build.
The updated Ethos-U driver v24.05 (introduced in #2626), contains a second
set of device functions for the U85. This can cause the linker to link the
functions of the U85, instead of the U55-U65, if they're both included in
the driver's sources files. Therefore a glob pattern can no longer be used
to collect the driver's source files in the Makefile.